+++ src/app/app.component.ts
@@ -11,18 +11,18 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MdInputModule, MdButtonModule, MdCheckboxModule, MdDatepickerModule } from '@angular/material';
 import { AppRoutingModule } from './app-routing.module';
 import { AppConfig } from './app.config';
-import { LoginComponent } from './login/login.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { GradingComponent } from './grading/grading.component';
-import { AssignmentsComponent } from './assignments/assignments.component';
-import { WeightingsComponent } from './weightings/weightings.component';
+import { LoginComponent } from './components/login/login.component';
+import { DashboardComponent } from './components/dashboard/dashboard.component';
+import { GradingComponent } from './components/grading/grading.component';
+import { AssignmentsComponent } from './components/assignments/assignments.component';
+import { WeightingsComponent } from './components/weightings/weightings.component';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
- 
+
 
  
   constructor() {
