+++ src/app/components/admin-marks-interface/admin-marks-interface.component.ts
@@ -2,7 +2,7 @@
     import {RedComponentComponent} from "../red-component/red-component.component";
     import {GridOptions} from "ag-grid/main";
     import {AgGridModule} from "ag-grid-angular";
-    
+
 @Component({
     selector: 'app-admin-marks-interface',
     templateUrl: './admin-marks-interface.component.html',
@@ -15,20 +15,20 @@
         i=0;
         constructor() {
             this.gridOptions = <GridOptions>{rowSelection: 'multiple'};
-    
+
             this.columnDefs = [
                 {headerName: "ID", field: "id",editable:true},
                 {headerName: "Total", field: "total", filter: 'number',editable:true,aggfunction:'sum'},
-                
+
             ];
-    
+
             this.rowData = [
                 {id: "u15062440", total: "0"},
                 {id: "u12345678", total: "0"},
                 {id: "u20982211", total: "0"}
             ]
         }
-    
+
         addColumn(colHeader)
         {
             var columnDefs = this.gridOptions.columnDefs;
@@ -53,17 +53,16 @@
             var transactions = {
                 add: [ {id: "x", total: "0"},]
             }
-            
+
             this.gridOptions.api.updateRowData(transactions);
         }
-        
+
         onGridReady(params) {
             params.api.sizeColumnsToFit();
         }
-    
+
         selectAllRows() {
             this.gridOptions.api.selectAll();
         }
     }
-    
-    
+
