+++ src/app/components/admin-query/admin-query.component.spec.ts
