+++ src/app/components/assignments/app.config.ts
