+++ src/app/components/assignments/assignments.component.scss
