+++ src/app/components/assignments/assignments.component.spec.ts
@@ -1,13 +1,13 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MdDatepickerModule,MdInputModule } from '@angular/material';
 import { AssignmentsComponent } from './assignments.component';
-import {UsersApi} from '../services/api-service/users-api/users-api.mock';
-import {User} from '../models/User';
+import {UsersApi} from '../../services/api-service/users-api/users-api.mock';
+import {User} from '../../models/User';
 import {Injectable} from '@angular/core';
 import { HttpModule } from '@angular/http';
-import {AssessmentsApi} from '../services/api-service/assessments-api/assessments-api.service';
+import {AssessmentsApi} from '../../services/api-service/assessments-api/assessments-api.service';
 
-import {Assessment} from '../models/Assessment';
+import {Assessment} from '../../models/Assessment';
 describe('AssignmentsComponent', () => {
   let component: AssignmentsComponent;
   let fixture: ComponentFixture<AssignmentsComponent>;
