+++ src/app/components/dashboard/app.config.ts
