+++ src/app/components/grading/grading.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import {UsersApi} from '../services/api-service/users-api/users-api.mock';
-import {User} from '../models/User';
+import {UsersApi} from '../../services/api-service/users-api/users-api.mock';
+import {User} from '../../models/User';
 import {Injectable} from '@angular/core';
-import {ApiService} from '../services/api-service/api.service';
+import {ApiService} from '../../services/api-service/api.service';
 import { HttpModule } from '@angular/http';
-import { AppConfig } from '../app.config';
+import { AppConfig } from '../../app.config';
 @Component({
   selector: 'app-grading',
   templateUrl: './grading.component.html',
@@ -30,22 +30,22 @@ export class GradingComponent implements OnInit {
 
  getModules()
   {
-  
+
     this.userService.getCurrentUser().subscribe(
     function(response) { this.currentUser=response},
     function(error) { console.log("Error happened" + error)},
-    function() 
+    function()
     {
       document.getElementById("sel1").innerHTML = "";
       for (var I = 0; I < this.currentUser.modules.length; I++)
       {
           var moduleList = "<option>" + this.currentUser.modules[I] + "</option>";
           document.getElementById("sel1").innerHTML += moduleList;
-          
+
       }
     });
-  
- 
+
+
 }
 
 }
