+++ src/app/components/leaderboard/app.config.ts
