+++ src/app/components/leaderboard/leaderboard.component.scss
