+++ src/app/components/red-component/red-component.component.html
