+++ src/app/components/red-component/red-component.component.spec.ts
