+++ src/app/components/student-dash/student-dash.component.ts
@@ -1,13 +1,13 @@
 import {Component, OnInit } from '@angular/core';
 import {MdCardModule} from '@angular/material';
-import {AuthService } from '../services/auth-service/auth.service';
-import {UsersApi} from '../services/api-service/users-api/users-api.mock';
-import {User} from '../models/User';
-import {Module} from '../models/Module';
+import {AuthService } from '../../services/auth-service/auth.service';
+import {UsersApi} from '../../services/api-service/users-api/users-api.mock';
+import {User} from '../../models/User';
+import {Module} from '../../models/Module';
 import {Injectable} from '@angular/core';
-import {ApiService} from '../services/api-service/api.service';
+import {ApiService} from '../../services/api-service/api.service';
 import {HttpModule } from '@angular/http';
-import {AppConfig } from '../app.config';
+import {AppConfig } from '../../app.config';
 @Component({
   selector: 'app-student-dash',
   templateUrl: './student-dash.component.html',
