+++ src/app/components/student-query/student-query.component.scss
