+++ src/app/components/weightings/weightings.component.html
