+++ src/app/components/weightings/weightings.component.scss
